+++ compiler/src/iree/compiler/DispatchCreation/BUILD.bazel
@@ -42,7 +42,6 @@ iree_compiler_cc_library(
         "HoistEncodingOps.cpp",
         "HoistUniformScalarCompute.cpp",
         "InsertTensorBarriers.cpp",
-        "LoopMappingUtils.cpp",
         "MaterializeDefaultWorkgroupCountRegion.cpp",
         "Passes.cpp",
         "PropagateEncodings.cpp",
@@ -56,7 +55,6 @@ iree_compiler_cc_library(
     ],
     hdrs = [
         "FusionUtils.h",
-        "LoopMappingUtils.h",
         "Passes.h",
     ],
     deps = [
