+++ compiler/src/iree/compiler/Utils/BUILD.bazel
@@ -39,6 +39,7 @@ iree_compiler_cc_library(
         "EquivalenceUtils.cpp",
         "FlatbufferUtils.cpp",
         "Indexing.cpp",
+        "LoopMappingUtils.cpp",
         "ModuleUtils.cpp",
         "OptionUtils.cpp",
         "PassUtils.cpp",
@@ -56,6 +57,7 @@ iree_compiler_cc_library(
         "FlatbufferUtils.h",
         "Indexing.h",
         "IntegerSet.h",
+        "LoopMappingUtils.h",
         "ModuleUtils.h",
         "OpVisitor.h",
         "OptionUtils.h",
@@ -70,6 +72,8 @@ iree_compiler_cc_library(
         "TracingUtils.h",
     ],
     deps = [
+        "//compiler/src/iree/compiler/Dialect/Encoding/IR",
+        "//compiler/src/iree/compiler/Dialect/LinalgExt/IR",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/internal/flatcc:building",
         "//runtime/src/iree/base/internal/flatcc:debugging",
@@ -84,9 +88,12 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:FuncDialect",
         "@llvm-project//mlir:FunctionInterfaces",
         "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:LinalgDialect",
+        "@llvm-project//mlir:LinalgUtils",
         "@llvm-project//mlir:Parser",
         "@llvm-project//mlir:Pass",
         "@llvm-project//mlir:Support",
+        "@llvm-project//mlir:TilingInterface",
         "@llvm-project//mlir:TransformUtils",
         "@llvm-project//mlir:Transforms",
     ],
